requirements.txt
Django==1.11
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
Markdown==2.6.11
PyJWT==1.6.4
pytz==2018.5
